Check commands return code value
Nagios doesn't consider check command result if its return code is
out of allowed bounds (0 <=> 3), displays the "Return code of XXX
is out of bounds" message and sets its status to "CRITICAL".
This change rectifies the return code value if it is out of the
allowed bounds so that users can actually see the check result error
message.